dump: Also validate sha256 sums when they are available
Since buster, sha256 checksums in package indices are mandatory
and md5 sums are only optional.
On security.debian.org md5 sums are already dropped for bullseye.
This breaks the fullpackagelist validation.

Validate sha256 sums when they are there. Make sure, that at least one
sum is validated in the validation.

Signed-off-by: Torben Hohn <torben.hohn@linutronix.de>
Reviewed-by: Holger Dengler <dengler@linutronix.de>