initvm: Remove cirrus vga device
Remove the complete video section to make use of the default,
which is provided on an existing graphics section.

Quoting from the libvirt website:
"""
The video element is the container for describing video devices.
For backwards compatibility, if no video is set but there is a graphics in
domain xml, then libvirt will add a default video according to the guest type.
"""

Link: https://libvirt.org/formatdomain.html
Signed-off-by: Bastian Germann <bage@linutronix.de>
Tested-by: Holger Dengler <holger@hdengler.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>