initvm: Prevent trying old kvm interpreter names
buster and later Debian versions just have a kvm symlink to qemu-system.
libvirt takes care of setting the right arguments so that KVM is enabled
with the default template. The qemu arguments are only kept for the initvm
creation.

There was an inconsistency in that amd64 had the full path to interpreter
and all other systems had only the executable name.
Make up for that by finding the path later via shutil.which().

The Debian package qemu-kvm, which used to have the kvm wrapper executable
is a virtual package. Virtual packages should only be used in Depends as
alternatives, so depend on the real package.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>