examples: Relicense to 0BSD
GPL-3.0-or-later is applicable to the example XML files but it is a bad
license for examples because it needs attribution and having the license
text available along with it. People copy and paste the examples as if
they were public domain, so apply the 0BSD license to them that does not
require attribution or keeping the license comment.

Add the license text to debian/copyright.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Viraj Shah <viraj.shah@linutronix.de>