schema: Add XSD's missing license info
The xsd.xml file was added without its applicable license which is not
referenced in the document itself. Replace the file with the exact version
provided by W3c (has some whitespace and > -> &gt; changes).

Add an SPDX header for it and the complete license info in debian/copyright.
The license requires that any modifications are logged with date, so mention
this modification in the comment itself.

Fixes: 31df3c46ea ("Add support for XML catalog")
Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Viraj Shah <viraj.shah@linutronix.de>