Replace copyright information with SPDX-FileCopyrightText
Use the SPDX-FileCopyrightText for copyright information instead of
free-form comments. While at it, replace the individual Linutronix
authors with Linutronix GmbH, which is the correct copyright assignment
for their contributions. Make sure that all of the removed names are listed
in the AUTHORS or THANKS files.

Add missing copyright info with the same format.
Fix some docs wording along the way.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>