hdimg: Allow additional arguments for grub-install
There might be use cases where grub-install needs additional arguments to
the ones that are currently used. Replace the erroneous schema documentation
to reflect this possibility and implement it in the actual grub-install calls.
The additional arguments are prepended to Elbe's so that they are overridden
by Elbe if duplicate arguments are supplied.

Extend one example to make use of the new feature.
While at it, fix too deep indentation.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Viraj Shah <viraj.shah@linutronix.de>