initvmaction: Run repodir before preprocess
repodir's logical place in the command pipeline is before preprocess
because preprocess will rely on any <repodir> element to be replaced.

Add the Repodir context manager at the right place and catch RepodirError.
The web servers will be terminated with the context manager.

The rest of the change is just indenting the inner context by two levels.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>