Fix bash completion support
When adding the bash completion support I renamed the bash-completion file
to be installed in the package where the executable is located.
Unfortunately, I tested the resulting Debian package build in the same bash
where the completion was already loaded and did not notice it does not work
that way.

As it is not Debian-related, move the completion file out of the debian
directory and reference it.

Fixes: 298ede3242 ("Add basic support for bash completion for elbe")
Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-By: Kaspar Giger <kg@kgmw.ch>