initvmaction: Let 'ensure' show the control error
'elbe control list_projects' is used in 'elbe ensure' when the initvm is
running to determine if the SOAP connection can be established.

Its output is suppressed which leads to users not knowing the cause of
not being able to connect, e.g. when the client was updated while the
server/initvm was not.

Make 'elbe initvm ensure' show the stderr of the underlying command.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>