Replace qemu -enable-kvm option
QEMU's -enable-kvm option maps to the old-style -machine accel which is
incompatible with the -accel option in current qemu versions. The -accel
option is implied by libvirt's 'hvm' machine type that is used with Elbe's
defaults.

Replace the -enable-kvm that ends up in libvirt so Elbe is compatible with
bookworm. The replacement is at least available since buster (oldest
supported release for the client).

Fixes: #347

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>