Eliminate most apt-key calls
apt-key is deprecated and calls to gpg executables, which is especially
bad in Elbe's target context. It makes sure to add the keyring in its
argument to apt's trusted keys and unarmors the keyring on the way.

Replace apt-key with unarmor_openpgp_keyring conversions and file writes
to /etc/apt/trusted.gpg.d/, which moves the gpg call out of the target
context.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: John Ogness <john.ogness@linutronix.de>