debian: Update debhelper compat to level 12
The elbe debian package is only built for buster and later, so update the
level to 12. The .service files are now installed with dh_installsystemd
instead of dh_installinit, which gets rid of one debian/rules override.

Prevent dh_installsystemd from enabling/starting elbe.service because its
environment refers to a non-existing file.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>