initvm: add an option for additional Debian archive key location
The option `keys_dir` provides a path to the additional key location.

The use case is when you create an initvm on, for example, Ubuntu
where the `debian-archive-keyring` package installs the keys to
`/usr/share/keyrings/`. As elbe looks in `/etc/apt/trusted.gpg.d`
for the keys, it ignores the ones in `/usr/share/keyrings/`.

Signed-off-by: Yegor Yefremov <yegorslists@googlemail.com>