rfs: Prevent cutting OpenPGP armor end line
Stripping the end line via :-1 can result in
"-----END PGP PUBLIC KEY BLOCK-----" being stripped depending on the raw-key
having a terminating line feed or not. Also, stripping the first line via 1:
is not necessary here. Get rid of the list range.

bullseye's gpg will complain about a wrong CRC when the end line is stripped,
which is misleading.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Ros Spolyak <rostokus@gmail.com>