Replace exit static exit codes with ascending numbers.
E.L.B.E. exits with error codes in different error
cases. The used error codes where 0,1,5,10,20 and 30.
Therfor it was hard to find the appropriate location
in the code where the error occurred. In order to
ease the debugging LX decided to spit out unique
error codes. After this change every sys.exit()
spits out a unique error code, which makes it easy
to identify the location where the error occurred.

Signed-off-by: Hartmut Eilers <eilers@linutronix.de>
Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>