Ignore DeprecationWarning from crypt
crypt is deprecated in Python 3.11 and scheduled for removal in
Python 3.13. It claims passlib as recommended replacement. Using passlib
result in an error:

DeprecationWarning: 'crypt' is deprecated and slated for removal in Python 3.13
    from crypt import crypt as _crypt

The passlib project is aware of that error, but has no solution for it
right now. Ignore the error until a better solution is available.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Eduard Krein <eduard.krein@linutronix.de>