sdk: Add missing run/debug components
A toolchain typically contains everything necessary to build, run,
and debug software. Although the ELBE SDK contains everything
necessary to build, it does not contain everything necessary to
run and debug the software. The missing components are:

ldconfig: To setup the dynamic loader cache on the target.

gdbserver: To perform remote debugging on the target.

glibc debug symbols: To debug the libraries provided by the
toolchain.

Add these 3 missing components.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Reviewed-by: Kurt Kanzenbach <kurt@linutronix.de>