Drop finetuning action "unit-tests"
The unit-tests finetunig action was initialy introduced to embed and
perform tests on the generated RFS into the E.L.B.E. receipe file.
This action relay on a python module, which is not packaged by Debian and
need manual interaction inside the initvm to install this module by hand.
Beside that the only test, that can be performed is a file exists tests.
Other tests were planed, but never implemented. On top of that, the used
python module pertube the autodiscovery of pytest or unittest. Therefore
drop this finetuning action.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>