elbepack: don't transform all warnings into exceptions
The calls to warnings.simplefilter() modify the global warning
configuration. Doing it at import-time and without
warnings.catch_warnings() means that this configuration is changed in
essentially unpredictable ways.

The call to `warnings.simplefilter('error')` transforms *all*
warnings for *all* modules into exceptions.
This breaks the initvm on bookworm as the six library used by the spyne
library triggers a warning on python3 which then crashes the
application [0].
The emitted warning however is completely harmless and even ignored in
the default configuration.

Instead change the configuration only in the narrow scope it is
necessary in.

[0] https://github.com/arskom/spyne/pull/704

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>