schema: Update xml.xsd schema location
The XML schema was enhanced by the xml:id attribute to allow unique ids
in XML elements. This attribute can be used by Xpointer to reference
XML elements and ease the usage of Xinclude.
Update the XML schema location and the local copy of the schema as
preparatory step to ease Xinclude usage.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>