elbepack: load xml files via importlib.resources
The environment files is effectively a global variable.
Also having hacks in the code for system-wide installation support is
ugly.

Instead transform the schema files into proper python package data that
can be loaded via importlib.resources.

The Debian packaging is adapted so the schemas are still installed to
the proper system-wide location and registered properly.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>