preferences: flexible pkg pinning options
This enables the use of apt package pinning options on package level
other than the codename, namely:
release-archive, release-component, release-label, release-name,
release-origin, release-version

For backwards compatibility, the former attribute 'pin' is mapped to
'release-name', but reported as deprecated.

The pinning options version, origin and release-* are not to be mixed,
as apt does not handle them separately. See also:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=868276

[thomas.weissschuh: add quotes around value of 'Pin: origin "VALUE"']
[thomas.weissschuh: use single quotes for string literals]

Signed-off-by: Franz Heger <franz.heger@kuka.com>
Reviewed-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>