README: add note about initvm create --devel
It is not immediately obvious that the ELBE version inside of the virtual
machine differs from the one running on the host, because the version inside
the VM is by default installed from debian.linutronix.de, while the one on
the host may come e.g. from a more recent git checkout.

This can result in a whole slew of issues, either due to the ELBE version
inside the virtual machine being incompatible with the one outside of it
or because one is trying to fix an issue but the changes do not show any
effect.

Add a hint about --devel to the README, where users using a git checkout
may find it first.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>
Reviewed-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>