elbevalidate: add package to validate elbe images
elbevalidate is a framework to introspect and validate images generated
by elbe. The framework is meant to be used in conjunction with a
unittesting library like pytest.
Images are not modified during testing, so tests are idempotent.

For now this package is experimental and only used for the
implementation of elbes own unittest.
In the future the goal is to stabilize the framework and expose it to
users of elbe.
Or the framework could be split into its own project to be used without
elbe.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>