elbepack: use devel initvm automatically when running from source tree
Elbe knows about two indicators for "development mode".
The first one is elbepack.version.is_devel which detects if elbe is
running from the source checkout.
The second one is "elbe initvm --devel" which needs to be specified by
the user to install the current source tree into the initvm.

Using --devel while is_devel == False was explicitly forbidden.
The inverse, using the non-devel initvm from a source checkout is not
guaranteed to work either and a source for confusion to users.

Instead automatically use the devel initvm when running from the source
tree.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>
Reviewed-by: Eduard Krein <eduard.krein@linutronix.de>