elbepack: cdroms: add source packages from debootstrap to cdrom
Binary packages installed during debootstrap are added to the binary
cdroms. As the same is not done for their source packages
"check-build cdrom" may fail.

Mismatches between debootstrap and apt installation can happen when the
security repositories contain updates, as those repositories are not
used by debootstrap.
For example currently this is the case with util-linux_2.36.1-8+deb11u1
and util-linux_2.36.1-8+deb11u2 in bullseye.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>
Reviewed-by: Eduard Krein <eduard.krein@linutronix.de>