elbepack: only provide resolv.conf during download step
Certain packages or finetunings expect to manage /etc/resolv.conf.
For example the resolvconf package contains a post-install script that
checks if no /etc/resolv.conf exists and if it does, it gets created as a
symlink to ../run/resolvconf/resolv.conf.

However the way elbe downloads and installs packages requires it to have
a working /etc/resolv.conf in the target.
This /etc/resolv.conf is copied from the initvm through the excursion
mechanism.

To enable both download with /etc/resolv.conf and installation without
/etc/resolv.conf, split the installation step into two phases, download
and installation.

Drawbacks:
* If a package installation script expects network connection this now
  fails. Arguably this would be bad practice in any case.

Closes: https://github.com/Linutronix/elbe/issues/274
Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>