elbepack: drop option --nesting
The --nesting option enables running the initvm within another vm.
This is achieved by telling libvirt to pass through the full host CPU to
the initvm.
Otherwise libvirt would refuse to start the VM as the virtual CPU would
not fulfill the requirements of its generic guest CPU.

Using a generic guest CPU is useful when migrating VMs between different
hosts. But in the context of an elbe initvm this is an edge case that
requires manual steps anyway.

More users are confused by the failure to start an initvm inside an
existing VM.

So default to the --nested behaviour and completely remove the option.

Generic CPU support may be added in the future as an optional feature.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>