elbepack: log: introduce shutdown protocol for async logger
The AsyncLogging object reads data from its pipe until all copies of
the writing end fd are closed.

The call to os.close(w) assumes that there are other copies of the write
fd around which is a very close coupling to its current callers, namely
the rpcaptcache manager and subprocess.Popen(close_fds=True, stderr=w).

These existing assumptions are non-obvious and brittle in the face of
code refactorings. Replace it with an obvious shutdown() method.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>