Elbe Docker container
This commit adds a Docker container for elbe which supports building
images without an initvm. It supports Debian and Ubuntu based images
and can built with a Debian or and Ubuntu base. Building images for
foreign architecutres is supported, if the binfmt_misc kernel module
is loaded in the host environment.

The command build_image is proided as helper for one-shot builds:

Usage: build_image
   [ -c | --build-sdk ]
   [ -r | --result-path ]
   -i | --image path/to/image/xml

Known issues:
-------------

- The initvm command doesn't work
- Building ISOs doesn't work

Signed-off-by: Thomas Irgang (thir820) <thomas.irgang@elektrobit.com>