elbepack: add check-image-list/check-script XML element
The contents of the XML element define a test script which can be used
to validate the result of an ELBE build.
The script will be executed by "elbe check-build img".

To make authorship easier the path to a script can be specified in the
"location" XML attribute and it will be inlined during preprocessing.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>