elbepack: shellhelper: introduce run() function
the run() function is a thin wrapper around subprocess.run(),
with the following changes:

* It defaults to check=True for more correctness
* It accepts shellper.ELBE_LOGGING for stdin and stdout to forward these
  through the elbe log

In contrast with the original shellhelper functions, this has
many advantages:

It is much more flexible for callers to achieve the exact behaviour they
require. Specifying stdout, stderr, cwd, check, etc; capturing output and
validating return codes are all available independently from each other.

Executing through a shell needs opt-in as that is more error-prone.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>