elbepack: finetuning: correctly parse mknod parameters
All mknod parameters are taken verbatim from a singular xml attribute.

While this is ugly as it ties the xml schema to the syntax of the mknod
command, there is not much to be done at this point.

Split the attribute value, so it ends up as multiple exec() arguments.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>