elbepack: treeutils: validate values of boolean attributes
According to the XML schema specification, boolean attributes have a
very limited set of valid values:

  3.2.2.1 Lexical representation
  An instance of a datatype that is defined as ·boolean· can have the
  following legal literals {true, false, 1, 0}.

  3.2.2.2 Canonical representation
  The canonical representation for boolean is the set of
  literals {true, false}.

Reflect this logic in the helper function.

While it won't make a difference at runtime, it is clearer to
understand.

Also return None if the attribute is missing.
The helper is split into its own toplevel function, as not all parts of
elbe use the etree objects but still need the functionality.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>