elbepack: treeutils: load local xml.xsd without catalog
While the Debian package properly sets up an XML catalog, the validation
also has to work without the catalog set up, for example during unittesting.

So far the xml.xsd was resolved through the internet in this case.
Resolve it to the local file which is already part of ELBE.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>