elbepack: initvm: introduce a dedicated initvm class
Instead of keeping the logic in the CLI handlers and forcing all initvm
interaction going through the CLI, introduce a dedicated class hierarchy
for initvms.
This enables easier interaction with the initvm throughout the codebase.
It removes the need to roundtrip all configuration parameters through
the CLI calls and make it easier to introduce new types of initvms with
different configurations.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>