docs: add link to documentation overview
The elbe website contains an index of documentation for all released
versions of elbe.
Add a link to each documentation page to that index.

While this link is static and does not automatically lead to the correct
page when switching versions it is simple and does not need any
maintenance or Javascript.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>