Revert "xmlprepocess: search and replace urls based on ELBE_MIRROR_SED"
The "mirror sed" feature is very specialised and
modifying the source XML based on environment variables is non-obvious.

If somebody needs this feature it can be easily implemented outside of
elbe, see contrib/replace-initvm-repository.py.

To simplify a future refactoring of the configuration handling,
drop the feature and its configuration.

This reverts commit 07179801530b15211fae77710a00348ada73d533.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>