elbepack: version: always mark devel version string
The development elbe version is distinct from the release version.
Make that clear in the version string.
To be compatible with python packaging version string requirements,
use the suffix '.dev0' instead of '-devel'.

Only the initial initvm setup needs access to the base version, as this
needs to install a Debian package which does not exist for devel
versions.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>
Reviewed-by: Eduard Krein<eduard.krein@linutronix.de>