elbepack: daemon: remove cherrypy usage
Only WSGI server from cherrypy is used.
And through private APIs at that.

Switch to the stdlib WSGI server.
While it is single-threaded, so is the elbe initvm itself.
In case multithreading is needed in the future it is still possible to
switch either to ThreadingHTTPServer or even a dedicated WSGI server
library.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>