elbepack: finetuning: run raw_cmd without a shell
In diet mode there very well may not be a shell available in the target
system during finetuning.
Run `<raw_cmd>` actions without a shell.
Historically a shell was used for `<raw_cmd>`, but that was the buildenv
one, leading to surprising results.

Removing the shell from `<raw_cmd>` now also makes the difference
between it and `<command>` clear.

Also align the schema documentation and tests to that effect.

Closes #408
Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>