elbepack: directories: introduce helper to execute an elbe subcommand
In contrast to the existing run_elbe() helper,
run_elbe_subcommand() executes the subcommand in the same
interpreter/process.
This has the advantage that all environment and stacktraces are
preserved.
For now this is useful for unittests but as soon as the exception-based
error handling is active it can be used throughout elbe.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>