elbepack: projectmanager: avoid cache indirection in project_is_busy()
The target project is already available, there is no need for the
indirection refetching it from the cache.

Avoid calling open_project() as that invokes the ElbeProject constructor
which has sideeffects.  These sideeffects break when called concurrently
with another constructor, but the whole point of project_is_busy() is to
be called concurrently.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>