elbepack: projectmanager: remove cache
The cache artificially limits each user to one active project at a time.
Retrieving projects by userid instead of their unique builddir, which is
always available has the potential to operate on an incorrect project.
It also introduces a global lock.

The only work that the cache optimizes are small data fetches from a
local database which don't need such a cache, and are cached in the ORM
anyways.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>