elbepack: log: simplify log handler housekeeping
The log module creates custom log handlers for each project.
These are then stored in a thread-local variable to clean them up when
not needed anymore.
This can be simplified by using a contextmanager, which coincidentally
is already used today.
Only the doctests can't use this because of syntax limitations in the
doctest module.
Push the added complexity into the doctests instead of the main logging
module where it affects everything.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>