elbepack: main: untangle run_elbe_subcommand() from main()
main() needs to do a few things which are unnecessary for
run_elbe_subcommand(). Instead of passing arguments around to disable
the unnecessary steps, separate the two functions.

This will also make it easier to add more initialization steps to main()
which would create issues to call from run_elbe_subcommand() like the
setup of the logging system.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>