elbepack: soapclient: automatically connect on service usage
Establish the connection when the service is used the first time,
instead of forcing the users to call .connect() manually first.
This also fixes the misleading error
'AttributeError: 'ElbeSoapClient' object has no attribute 'service'
which would happen in this case as .service is only set in .connect().

Keep the .connect() method so users still can force a connection at a
specific point.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>