elbepack: pkgutils: handle clearsigned dscs
.dsc files can be clearsigned which means they have the PGP signature in
the file itself. apt_pkg.TagFile() is unaware of this.
The previous elbepak.pkgutils.get_dsc_size() worked by chance as
normally the PGP header and trailer are separated by an empty line from
the actual tag contents and get_dsc_size() always skipped past the
trailer. There are however .dsc files which do not have a separating
newline in front of the trailer, making TagFile() think that the trailer
belongs to the tag section, breaking the parsing.
One of those is dsc files is now in debian-security:
ghostscript_10.0.0~dfsg-11+deb12u5.dsc

Switch over to apt.debfile.DscSrcPackage() which correctly strips the
PGP signature before parsing the file.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>