elbepack: log: work around multiprocessing bug
Old versions of Python have a bug that is triggered by our usage of the
multiprocessing module.
This bug is already fixed but the version of Python in Debian bullseye
does not and will not contain this patch.

Work around this by monkeypatching the affected stdlib function.
While is is obviously a hack it should be fine as it only replicates an
actually released bugfix.
While the value of 'manager_owned' should be passed through instead of
discarded, this seems to work well enough for the usecase in elbe.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>