elbepack: use source name of pkgs with on_src_cd attribute
Resolve the corresponding source name of the binary packages
where the XML attribute on_src_cd="False" is set.
Until now the binary name was used, which doesn't work when
the source name differs to the binary name of the package.
The forbidden_packages list, containing the binary package names
will be resolved and the corresponding source package names
are stored in the forbidden_src_packages list which is then used
for the comparison inside the add_source_pkg function.

Signed-off-by: Martin Hoehn <martin.hoehn@mail.de>