elbepack: aptpkgutils: reduce scope of apt import
We want to use elbepack.aptpkgutils also in environments which do not
have python-apt available. For example to use "elbe cyclonedx-sbom".
Reduce the scope of the imports to make that possible.

This is slightly iffy and the logic should be split over multiple
modules, which either depend on python-apt or don't.
However the internals of aptpkgutils are fairly intertwined and
splitting it up would make it only more ugly.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>