elbepack: soap: use urllib to download files
All project files are available through the repo http server.
Using that instead of the custom SOAP bytestream is much faster and
requires less memory and code.

This bypasses the permission check, but the repo server is available
anyways, so the files are already reachable by anyone.
The user authentication is also going be removed anyways.

Closes #418
Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>