elbepack: reduce memory usage of xz compressor
Users reported that the OOM killer would trigger inside the initvm due
to high xz memory usage.
Reduce that memory usage.
It will increase the compression time, but that's preferrable to
inscrutinable errors.

A plan for the next larger release is also to increase the total memory
of the initvm, as the current value of 1GiB is very small.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>