docs: elbe-initvm: fix order of options
Most options now have to be specified after the "elbe initvm ..."
sub-subcommand. They have been moved there because not all options are
valid for all sub-subcommands.
The documentation was not adapted for this.

Unfortunately the XML options still need to be specified before.
That is an oversight and will be corrected at some point, but until then
make sure the documentation matches reality.

Closes: https://github.com/Linutronix/elbe/issues/414
Closes: https://github.com/Linutronix/elbe/issues/421
Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>