elbepack: add /lib64 to target sysroot
The libc.so is a linker script having `AS_NEEDED ( /lib64/ld-linux-x86-64.so.2
)`. While on the sysroot host thix gets fixed by the relocation script, on the
target it stays untouched. However, when a user links the C library by directly
mention the path to ``sysroot/target/usr/lib/x86_64-linux-gnu/libc.so` it gets
the error that `/lib64/ld-linux-x86-64.so.2` can't be found. Or worse, it links
using the user environment. On the other hand, this does not happen when using
`-lc` or letting `gcc` pick the right libc because it fetches the one from the
sysroot host. This is not a problem on x86_64 because both host and target
sysroot share the same binaries.

This patch adds the `/lib64` symbolic link to the sysroot target.

Signed-off-by: Federico Vaga <federico.vaga@vaga.pv.it>
[thomas: fix list syntax whitespace]
Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>