elbepack: elbeproject: log used elbe version
Knowing the used version of ELBE is useful when debugging issues.
While version is already part of elbe-report.txt, that is not shown to
the user and in case of failed builds needs to be downloaded manually.
Print the used version to the default output where it is easily usable.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>