elbepack: cdroms: log encountered exceptions
If the download fails due to an exception that should be reported to the
user so they can handle or fix the issue.
As all logging happens while an exception is active, logging.exception()
provides this information conveniently.
The same exception logging already exists in elbepack.finetuning and
elbepack.commands.fetch_initvm_pkgs.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>