elbepack: finetuning: fix file base64 encoding
standard_b64decode() always returns bytes, but the methods from the
elbepack.filesystem.Filestem() that these are passed to expect a string.
This probably worked on Python 2 but it doesn't today.
While it would be nice to use arbitrary binary data with <file>,
that is a feature change for the future.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>