elbepack: imgutils: protect against race between losetup() and udev
After creation of the loop device through losetup, udev reconfigures it.
Partitions are rescanned, removing and then readding their device nodes.
This introduces a race window where ELBE tries to access the non-existent
partition device nodes after creating the loop device.
The involvment of udev is unnecessary anyways as devtmpfs will properly create
the partition device nodes on its won.

Use a file lock to synchronize the accesses between udev and ELBE.
The lock will be dropped automatically when the filedescriptor is closed
at the end of the contextmanager.
It is not defined if ELBE runs before udev or the other way round, but both
cases are fine for our usecase.

Reported-by: John Ogness <jogness@linutronix.de>
Link: https://systemd.io/BLOCK_DEVICE_LOCKING
Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>
Reviewed-by: John Ogness <jogness@linutronix.de>