elbepack: hdimg: clarify glob match error
If there is no match for a glob pattern, the error message does not show the
pattern that caused the error. This makes it unnecessary harder to debug the
issue.
Add args to the error message and clarify the error message for different
cases to make it easier to debug such issues.

Signed-off-by: Vasilij Strassheim <vasilij.strassheim@linutronix.de>
Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>