elbepack: imgutils: properly wait for all udev events in losetup()
grub-install called inside the target expects the /dev/disk/by-uuid/ links to
be set up properly. Otherwise plain device names are used, which in the case of
loop devices does not work at all.
Depending on the timing, the current code could block execution of udev and
therefore prevent the creation of the symlinks.

Reported-by: Thomas Bonnefille <thomas.bonnefille@bootlin.com>
Closes: https://github.com/Linutronix/elbe/issues/431
Fixes: c5e73aadb421 ("elbepack: imgutils: protect against race between losetup() and udev")
Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>