Revert "elbepack: asyncworker: use multiprocessing as execution backend"
The different multiprocessing managers used within ELBE seem to interact badly
in some cases and may lead to errors like the following:

```
  File "/usr/lib/python3.11/logging/__init__.py", line 978, in handle
    self.emit(record)
  File "/var/cache/elbe/devel/elbepack/log.py", line 79, in emit
    self.Q.append(self.format(record))
  File "<string>", line 2, in append
  File "/usr/lib/python3.11/multiprocessing/managers.py", line 821, in _callmethod
    conn.send((self._id, methodname, args, kwds))
  File "/usr/lib/python3.11/multiprocessing/connection.py", line 205, in send
    self._send_bytes(_ForkingPickler.dumps(obj))
  File "/usr/lib/python3.11/multiprocessing/connection.py", line 410, in _send_bytes
    self._send(header + buf)
  File "/usr/lib/python3.11/multiprocessing/connection.py", line 367, in _send
    n = write(self._handle, buf)
        ^^^^^^^^^^^^^^^^^^^^^^^^
OSError: [Errno 9] Bad file descriptor
```

Depending on the specific ELBE usage pattern these exception may trigger very
infrequently or every time.
The usage of multiprocessing in the asyncworker framework was a preparation for
parallel execution of async workers.
As that followup work is not finished yet, fall back to the old threading based
implementation and figure out the root cause of the issue afterwards.

This reverts commit 8850f7ec5482 ("elbepack: asyncworker: use multiprocessing as execution backend")

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>