contrib: dockerfile: Update to bookworm and drop libvirt compatibility
The current container is running debian bullseye and is starting
systemd to run libvirtd in background. This container is broken since
some times.
This commit updates the dockerfile to run interactively, libvirt isn't
supported in this version.
There are two launch options, start to run in production which will
use the packaged version of Elbe and start-devel which will use the
version of Elbe currently being used by the repository.

Signed-off-by: Thomas Bonnefille <thomas.bonnefille@bootlin.com>