elbepack: rpcaptcache: enable pkgProblemResolver debug output
Apt has a dependency resolution stage called "pkgProblemResolver". It's
responsible to make a decision when two packages want to be installed,
but only one of them can be installed. With the config option enabled
here we can watch the decision process in elbe's build log. It is
essential to figure out the root cause when apt bails out after a chain
of such problems.

Since this increases the log load only in case of complex dependency
problems and is silent otherwise, we can enable it unconditionally.

See also apt.conf(5).

Signed-off-by: Tobias Deiminger <tobias.deiminger@linutronix.de>
Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>