elbepack: always provide default packages in the SDK
When the user does not provide any `<hostsdk-pkg-list>`, ELBE adds by default a
few packages. However, these are not added when the user adds a package to
`<hostsdk-pkg-list>`. In this second case, the user needs to explicitly add back
the default packages.

If the philosophy is that users always get the default set of packages, then
this patch fixes the problem.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>