elbepack: treeutils: enable decompression on lxml >= 6.0
Since version 6.0 lxml does not decompress its input anymore by default.
However ELBE relies on this feature and as ELBE expects its input to be
trusted the security reasoning does not apply.

As we need to stay compatible with older versions of lxml which do not
know this parameter add some conditional logic. Due to the
implementation of lxml, inspect.signature() does not work here.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>