elbepack: hdimg: use a list of options for mkfs
When the invocation of mkfs.ubifs stopped using a shell, the
specification of multiple custom mkfs arguments from XML was broken.
Arguments in the XML are represented as a single string which then was
not split before passing it on to mkfs.ubifs.

Split the arguments directly after the read from the XML.

Also rename the member to make it more clear that it contains a list.

Reported-by: Ralf Schlatterbeck <rsc@runtux.com>
Link: https://lists.linutronix.de/elbe-users/20251112133331.svhat62vj5c5z6nj@runtux.com/
Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>