elbepack: main: fix --help output of subcommands
The toplevel argument parser does not have access to the full subparsers
with all their argument definitions. Doing so would require the import
of all subcommand modules, sacrificing performance. Instead the toplevel
parser only knows the subcommands name and import them modules on demand
if one is to be executed.
This works generally well except that currently an invocation of 'elbe
<subcommand> --help' only ever shows an empty help text. This is because
interpretation of '--help' is done by the subparser object registered to
the toplevel parser. This subparser is only a dummy and does not know
about the actual arguments known to the command module.

Inhibit the creation of the implicit '--help' action on the dummy parser
so it instead gets forwarded to the command module and handled there.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>